+++ src/pages/Search/components/RollingBanner/RollingBanner.module.scss
@@ -24,6 +24,7 @@ $default: #c4c4c4;
     height: 100%;
     width: 38px;
     pointer-events: none;
+
     // z-index: 1;
   }
