+++ src/pages/Search/components/SearchInput/SearchInput.module.scss
@@ -14,7 +14,7 @@
   margin-top: 34px;
   border: none;
   border-radius: 31px;
-  background-color: #FFFFFF;
+  background-color: #ffffff;
   box-shadow: 2px 3px 12px 1px rgba(0 0 0 / 10%);
   z-index: 2;
