+++ src/pages/Search/components/Sentence/index.tsx
