+++ src/pages/Search/components/Suggestions/Suggestions.module.scss
@@ -13,8 +13,7 @@
 }
 
 .autocomplete {
-
-  &__box{
+  &__box {
     display: flex;
     height: 60px;
     padding-top: 38px;
